fixed warning
git-svn-id: svn+ssh://svn.code.sf.net/p/jillion/code/trunk@1679 d1f56cec-0f3b-4699-95c7-aedf87c10f2a

+++ src/org/jcvi/jillion/trace/fastq/FastqFileParser.java
@@ -110,7 +110,7 @@ public abstract class FastqFileParser {
         //now parse the qualities
         int expectedQualities =  (int)sequence.getLength();
 		 
-        StringBuilder qualityBuilder = new StringBuilder((int)expectedQualities);
+        StringBuilder qualityBuilder = new StringBuilder(expectedQualities);
         
     	while(qualityBuilder.length() < expectedQualities){
     		line = parser.nextLine();
